requirements.txt
Django==5.1
psycopg2-binary==2.9.9
python-decouple==3.8 # para acceder al archivo .env
pillow==10.4.0 # para gestionar imagenes
django-widget-tweaks==1.5.0
django-mathfilters==1.0.0
djangorestframework==3.15.2

# paquetes para el algoritmo
opencv-python==4.10.0.84
numpy==1.26.4
protobuf==3.20
deepface==0.0.93
keras==2.8
tensorflow[and-cuda]==2.8
tensorflow-gpu==2.8
tensorboard==2.8

# paquetes para barra de progreso
celery==5.4.0
celery-progress==0.4 # progress bar
redis==5.0.8
